add the library features
