+++ wombat/core/__init__.py
@@ -4,7 +4,7 @@ Initializes the simulation modules, classes, and functions.
 isort:skip_file
 """
 
-
+import library
 from .data_classes import (
     Maintenance,
     Failure,
