+++ wombat/utilities/__init__.py
@@ -6,6 +6,5 @@ from .utilities import (
     convert_dt_to_hours,
     format_events_log_message,
     hours_until_future_hour,
-    load_yaml,
     setup_logger,
 )
