+++ wombat/utilities/utilities.py
@@ -3,50 +3,10 @@
 
 import datetime  # type: ignore
 import logging  # type: ignore
-import os  # type: ignore
-import re  # type: ignore
-from pathlib import Path
 from typing import Any, Callable, Union  # type: ignore
 
 import numpy as np  # type: ignore
 import pandas as pd  # type: ignore
-import yaml  # type: ignore
-
-
-# YAML loader that is able to read scientific notation
-custom_loader = yaml.SafeLoader
-custom_loader.add_implicit_resolver(
-    "tag:yaml.org,2002:float",
-    re.compile(
-        """^(?:
-     [-+]?(?:[0-9][0-9_]*)\\.[0-9_]*(?:[eE][-+]?[0-9]+)?
-    |[-+]?(?:[0-9][0-9_]*)(?:[eE][-+]?[0-9]+)
-    |\\.[0-9_]+(?:[eE][-+][0-9]+)?
-    |[-+]?[0-9][0-9_]*(?::[0-5]?[0-9])+\\.[0-9_]*
-    |[-+]?\\.(?:inf|Inf|INF)
-    |\\.(?:nan|NaN|NAN))$""",
-        re.X,
-    ),
-    list("-+0123456789."),
-)
-
-
-def load_yaml(path: Union[str, Path], fname: str) -> Any:
-    """Loads and returns the contents of the YAML file.
-
-    Parameters
-    ----------
-    path : Union[str, Path]
-        Path to the file to be loaded.
-    fname : str
-        Name of the file (ending in .yaml) to be loaded.
-
-    Returns
-    -------
-    Any
-        Whatever content is in the YAML file.
-    """
-    return yaml.load(open(os.path.join(path, fname), "r"), Loader=custom_loader)
 
 
 def convert_dt_to_hours(diff: datetime.timedelta) -> float:
